Avoid leaking goroutines when cleaning up instances
The code before this commit starts a goroutine every time
instanceCleaner.Run() is called to print the errors from
instanceCleaner.fetchInstancesToDelete(…), but this goroutine will never
finish. By closing the channel when the method is done, the goroutine is
cleanly stopped.

Also removed sending nil to the channels, since this had no additional
purpose when the channels are closed.